991 Move setDocketEntryFormForEditAction
